mcp-memu/requirements.txt
mcp>=1.0.0
fastmcp
python-dotenv
requests
flask
